+++ src/UglyToad.PdfPig/Content/Letter.cs
@@ -35,7 +35,7 @@
         /// <summary>
         /// The width occupied by the character within the PDF content.
         /// </summary>
-        public decimal Width { get; }
+        public double Width { get; }
 
         /// <summary>
         /// Position of the bounding box for the glyph, this is the box surrounding the visible glyph as it appears on the page.
@@ -47,7 +47,7 @@
         /// <summary>
         /// Size as defined in the PDF file. This is not equivalent to font size in points but is relative to other font sizes on the page.
         /// </summary>
-        public decimal FontSize { get; }
+        public double FontSize { get; }
 
         /// <summary>
         /// The name of the font.
@@ -57,12 +57,12 @@
         /// <summary>
         /// The size of the font in points. This is not ready for public consumption as the calculation is incorrect.
         /// </summary>
-        internal decimal PointSize { get; }
+        internal double PointSize { get; }
 
         /// <summary>
         /// Create a new letter to represent some text drawn by the Tj operator.
         /// </summary>
-        internal Letter(string value, PdfRectangle glyphRectangle, PdfPoint startBaseLine, PdfPoint endBaseLine, decimal width, decimal fontSize, string fontName, decimal pointSize)
+        internal Letter(string value, PdfRectangle glyphRectangle, PdfPoint startBaseLine, PdfPoint endBaseLine, double width, double fontSize, string fontName, double pointSize)
         {
             Value = value;
             GlyphRectangle = glyphRectangle;
@@ -85,7 +85,7 @@
 
         private TextDirection GetTextDirection()
         {
-            if (System.Math.Abs(StartBaseLine.Y - EndBaseLine.Y) < 10e-5m)
+            if (System.Math.Abs(StartBaseLine.Y - EndBaseLine.Y) < 10e-5)
             {
                 if (StartBaseLine.X > EndBaseLine.X)
                 {
@@ -93,7 +93,7 @@
                 }
                 return TextDirection.Horizontal;
             }
-            else if (System.Math.Abs(StartBaseLine.X - EndBaseLine.X) < 10e-5m)
+            else if (System.Math.Abs(StartBaseLine.X - EndBaseLine.X) < 10e-5)
             {
                 if (StartBaseLine.Y > EndBaseLine.Y)
                 {
