+++ src/UglyToad.PdfPig/Fonts/Type1/CharStrings/Commands/Arithmetic/CallSubrCommand.cs
@@ -23,7 +23,7 @@
 
             foreach (var command in subroutine.Commands)
             {
-                command.Match(x => context.Stack.Push(x),
+                command.Match(x => context.Stack.Push((double)x),
                     x => x.Run(context));
             }
         }
