+++ src/UglyToad.PdfPig/Fonts/Type1/CharStrings/Type1CharStrings.cs
@@ -93,7 +93,7 @@
 
             foreach (var command in sequence.Commands)
             {
-                command.Match(x => context.Stack.Push(x),
+                command.Match(x => context.Stack.Push((double)x),
                     x => x.Run(context));
             }
