+++ src/UglyToad.PdfPig/Graphics/Operations/PathConstruction/AppendEndControlPointBezierCurve.cs
@@ -55,13 +55,13 @@
         /// <inheritdoc />
         public void Write(Stream stream)
         {
-            stream.WriteDecimal(ControlPoint1.X);
+            stream.WriteDecimal((decimal)ControlPoint1.X);
             stream.WriteWhiteSpace();
-            stream.WriteDecimal(ControlPoint1.Y);
+            stream.WriteDecimal((decimal)ControlPoint1.Y);
             stream.WriteWhiteSpace();
-            stream.WriteDecimal(End.X);
+            stream.WriteDecimal((decimal)End.X);
             stream.WriteWhiteSpace();
-            stream.WriteDecimal(End.Y);
+            stream.WriteDecimal((decimal)End.Y);
             stream.WriteWhiteSpace();
             stream.WriteText(Symbol);
             stream.WriteNewLine();
